requirements.txt
Django==1.6.2
dj-database-url==0.2.2
django-taggit==0.11.2
psycopg2==2.5.2
sorl-thumbnail==11.12
argparse==1.2.1
django-rq==0.6.1
httplib2==0.8
python-dateutil==2.2
pytz==2013.9
redis==2.9.1
rq==0.3.13
six==1.5.2
times==0.6.2
wsgiref==0.1.2

